requirements.txt
python-binance
python-dotenv
rich
pandas
numpy
ta
pytest
streamlit

